Apply review suggestions: * Don't give impression that set_user_env can write environment variables for general purposes on Unix systems. * More descriptive variable name in get_user_environment_variables * Move restore_user_env to spyder/app/tests/conftest.py to avoid Spyder imports in top-level conftest.py
