+++ conftest.py
@@ -11,7 +11,6 @@ NOTE: DO NOT add fixtures here. It could generate problems with
       QtAwesome being called before a QApplication is created.
 """
 
-# Standard libary imports
 import os
 import os.path as osp
 import re
@@ -23,11 +22,6 @@ os.environ['SPYDER_PYTEST'] = 'True'
 
 # ---- Pytest adjustments
 import pytest
-
-# Spyder imports
-from spyder.config.base import running_in_ci
-from spyder.utils.environ import (get_user_env, set_user_env,
-                                  amend_user_shell_init)
 
 
 def pytest_addoption(parser):
@@ -142,20 +136,3 @@ def reset_conf_before_test():
 
     CONF.set('completions', 'provider_configuration', provider_configurations,
              notification=False)
-
-
-@pytest.fixture
-def restore_user_env():
-    """Set user environment variables and restore upon test exit"""
-    if not running_in_ci():
-        pytest.skip("Skipped because not in CI.")
-
-    if os.name == "nt":
-        orig_env = get_user_env()
-
-    yield
-
-    if os.name == "nt":
-        set_user_env(orig_env)
-    else:
-        amend_user_shell_init(restore=True)
